// src/client/models/TheMemberModifyViewModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TheMemberModifyViewModel = {
    ID?: string;
    MemberName: string;
    TeamId?: string;
    TitleId?: string;
    TitleName?: string;
    Salary?: string;
    Password?: string;
    Email?: string;
    WindowsIdentity?: string;
    Enabled?: boolean;
    Role?: number;
    Tags?: Array<string>;
    ImageData?: string;
};