// src/client/models/TheMemberViewModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
import type { DefinitionMemberViewModel } from './DefinitionMemberViewModel';
export type TheMemberViewModel = {
    Abbreviation?: string;
    RoleName?: string;
    TeamName?: string;
    ImageUrl?: string;
    ManagementTitleId?: string;
    EmployeeTitleId?: string;
    EmployeeTitles?: Array<DefinitionMemberViewModel>;
    ManagerTitles?: Array<DefinitionMemberViewModel>;
    TagsDisplay?: string;
    EnabledDisplay?: string;
    ID?: string;
    MemberName: string;
    TeamId?: string;
    TitleId?: string;
    TitleName?: string;
    Salary?: string;
    Password?: string;
    Email?: string;
    WindowsIdentity?: string;
    Enabled?: boolean;
    Role?: number;
    Tags?: Array<string>;
    ImageData?: string;
};